// front/src/app/models/gost.ts
export class gost{
    username: string = "";
    password: string = "";
    securityQuestion: string = "";
    securityAnswer: string = "";
    name: string = "";
    surname: string = "";
    gender: string = "";
    address: string = "";
    phoneNumber: string = "";
    email: string = "";
    pictureUrl: string = "";
    creditCardNumber: string = "";
    status: string = "";
}